Fix #129: Add username and password arguments to WFS class constructors
This changeset allows passing low-level HTTP authentication details
to high level WebFeatureService API.
